cleanup container adaption code
